requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0
openai>=1.3.0
tiktoken>=0.5.0
aiofiles>=23.2.0
numpy>=1.24.0,<2.0.0
